pyproject.toml
[tool.poetry]
name = "random-facts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
ffmpeg-python = "^0.2.0"
opencv-python = "^4.11.0.86"
diffusers = "^0.32.2"
jupyterlab = "^4.3.4"
ipykernel = "^6.29.5"
torch = "^2.5.1"
python-dotenv = "^1.0.1"
langchain = "^0.3.14"
langchain-mistralai = "^0.2.4"
langchain-community = "^0.3.14"
transformers = "^4.48.0"
accelerate = "^1.3.0"
mistralai = "^1.3.1"
langgraph = "^0.2.64"
lmnt = "^1.1.7"
wave = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
